// src/enums/emergency.ts
export enum EmergencyPlanScenario {
  NATURAL_DISASTER = 'NATURAL_DISASTER',
  INFRASTRUCTURE_FAILURES = 'INFRASTRUCTURE_FAILURES',
  SECURITY_INCIDENTS = 'SECURITY_INCIDENTS',
  HEALTH_EMERGENCIES = 'HEALTH_EMERGENCIES',
  HAZARDOUS_MATERIAL_INCIDENTS = 'HAZARDOUS_MATERIAL_INCIDENTS',
  FIRE_EMERGENCY = 'FIRE_EMERGENCY',
}

export enum Procedures {
  SAFETY = 'SAFETY',
  COMMUNICATIONS = 'COMMUNICATIONS',
  RESPONSE = 'RESPONSE',
  RECOVERY = 'RECOVERY',
  MONITORING_AND_EVALUATION = 'MONITORING_AND_EVALUATION',
  COMPLIANCE_AND_REGULATION = 'COMPLIANCE_AND_REGULATION',
}

export enum KeysContact {
  BUILDING = 'BUILDING',
  CONTACT = 'CONTACT',
  EMERGENCY_PLAN = 'EMERGENCY_PLAN',
  REGION = 'REGION',
}

export enum EmergencyReportGetPlanEnum {
  BY_SCENARIO = 'BY_SCENARIO',
  BY_AFFECTED_AREA = 'BY_AFFECTED_AREA',
  INFRASTRUCTURE_FAILURES = 'INFRASTRUCTURE_FAILURES',
}